refactor: unify note create/edit pages into shared NoteEditorForm (fixes #234)
Extract NoteEditorForm component (fields, char count, template selector,
AI suggestions panel, zero-links modal, auth gate, autocomplete fetch)
from the near-duplicate create and edit note pages. Routes are now thin
wrappers: create keeps draft/atomicity/post-save flow, detail keeps
view mode and link refresh.

Also fixes dead Tailwind classNames in the detail page's edit mode
(Tailwind is not installed, so the edit grid and error box were
unstyled) and brings edit mode up to create-page parity (char count,
form hints, theme-aware button colors).

Docs: mark shipped AI features in NOTES.md and preserve the #15
toggle-modes table before closing that epic.
