+++ frontend/src/app/knowledge-base/notes/[id]/page.module.scss
@@ -326,61 +326,16 @@
   }
 }
 
-// ===== EDIT MODE =====
-.saveActions {
-  @include flex-between;
-  margin-bottom: $spacing-4;
-  padding: $spacing-3;
-  background-color: $orange-50; // Warm background for primary action area
+// Error message (view mode)
+.errorBox {
+  padding: $spacing-4;
   border-radius: $border-radius-md;
-  border: $border-width-thin solid $orange-200;
+  border: $border-width-thin solid rgba(220, 38, 38, 0.3);
+  background-color: rgba(254, 226, 226, 0.8);
+  margin-bottom: $spacing-4;
 
-  .actionsLeft {
-    @include flex-start;
-    gap: $spacing-2;
-
-    .saveButton {
-      padding: $spacing-2 $spacing-4;
-      background-color: $orange-600; // Warm primary action
-      color: white;
-      border: none;
-      border-radius: $border-radius-button;
-      font-weight: $font-weight-semibold;
-      cursor: pointer;
-      transition: all 0.2s ease;
-
-      &:hover:not(:disabled) {
-        background-color: $orange-700;
-      }
-
-      &:disabled {
-        cursor: not-allowed;
-        opacity: 0.5;
-      }
-    }
-
-    .cancelButton {
-      padding: $spacing-2 $spacing-4;
-      background-color: transparent;
-      color: $color-text-primary;
-      border: $border-width-thin solid $color-border;
-      border-radius: $border-radius-button;
-      cursor: pointer;
-      transition: all 0.2s ease;
-
-      &:hover:not(:disabled) {
-        background-color: $color-bg-secondary;
-      }
-
-      &:disabled {
-        cursor: not-allowed;
-        opacity: 0.5;
-      }
-    }
-  }
-
-  .helpText {
-    @include text-tertiary;
+  .errorMessage {
+    color: $color-error-border;
   }
 }
 
@@ -402,172 +357,4 @@
   white-space: pre-wrap;
   color: $color-text-secondary;
   line-height: $line-height-relaxed;
-}
-
-// ===== FORM FIELDS (Edit Mode) =====
-.formLabel {
-  display: block;
-  margin-bottom: $spacing-1;
-  font-size: $font-size-sm;
-  font-weight: $font-weight-medium;
-  color: $color-text-secondary;
-}
-
-.formInput {
-  width: 100%;
-  padding: $spacing-2 $spacing-3;
-  border: $border-default;
-  border-radius: $border-radius-button;
-  background-color: $white;
-  transition: all 0.2s ease;
-
-  &:focus {
-    outline: none;
-    border-color: $color-input-border-focus;
-    box-shadow: $shadow-focus-primary;
-  }
-
-  &::placeholder {
-    color: $color-input-placeholder;
-  }
-}
-
-// ===== REFERENCE TOGGLE =====
-.referenceToggle {
-  margin-top: $spacing-2;
-}
-
-.checkboxLabel {
-  display: inline-flex;
-  align-items: center;
-  gap: $spacing-2;
-  cursor: pointer;
-
-  .checkbox {
-    width: 16px;
-    height: 16px;
-    accent-color: $color-link-default;
-  }
-
-  .checkboxText {
-    font-size: $font-size-sm;
-    font-weight: $font-weight-medium;
-    color: $color-text-primary;
-  }
-}
-
-.formHint {
-  margin-top: $spacing-1;
-  font-size: $font-size-xs;
-  color: $color-text-tertiary;
-}
-
-.suggestLinksButton {
-  padding: $spacing-6 $spacing-6;
-  border-radius: $border-radius-button;
-  border: $border-width-thin solid $color-border-strong;
-  background-color: $color-surface-tertiary;
-  color: $color-text-secondary;
-  font-weight: $font-weight-medium;
-  cursor: pointer;
-  transition: all 0.2s ease;
-
-  &:hover {
-    background-color: $color-surface-hover;
-  }
-}
-
-.sidebarPanel {
-  padding: $spacing-6;
-  border-radius: $border-radius-md;
-  border: $border-width-thin solid $color-border-subtle;
-  background-color: $white;
-}
-
-// ===== MODALS =====
-.modalOverlay {
-  position: fixed;
-  inset: 0;
-  z-index: 50;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: rgba($black, 0.5);
-}
-
-.modalContent {
-  max-width: 28rem;
-  padding: $spacing-6;
-  border-radius: $border-radius-lg;
-  background-color: $white;
-  box-shadow: $shadow-card-lg;
-}
-
-.modalTitle {
-  margin-bottom: $spacing-3;
-  font-size: $font-size-lg;
-  font-weight: $font-weight-semibold;
-  color: $color-text-primary;
-}
-
-.modalText {
-  margin-bottom: $spacing-4;
-  color: $color-text-secondary;
-}
-
-.modalSubtext {
-  margin-bottom: $spacing-4;
-  font-size: $font-size-sm;
-  color: $color-text-tertiary;
-}
-
-.modalList {
-  margin-bottom: $spacing-6;
-  padding-left: $spacing-5;
-  list-style-type: disc;
-  list-style-position: inside;
-
-  li {
-    margin-bottom: $spacing-1;
-    font-size: $font-size-sm;
-    color: $color-text-secondary;
-  }
-}
-
-.modalActions {
-  display: flex;
-  gap: $spacing-3;
-
-  .modalButton {
-    flex: 1;
-    padding: $spacing-2 $spacing-4;
-    border-radius: $border-radius-button;
-    font-weight: $font-weight-medium;
-    cursor: pointer;
-    transition: all 0.2s ease;
-    border: none;
-
-    &.primary {
-      background-color: $color-link-default;
-      color: $white;
-
-      &:hover {
-        background-color: $color-text-secondary;
-      }
-    }
-
-    &.secondary {
-      background-color: transparent;
-      color: $color-text-secondary;
-      border: $border-width-thin solid $color-border;
-
-      &:hover {
-        background-color: $color-bg-secondary;
-      }
-
-      &:disabled {
-        opacity: 0.5;
-      }
-    }
-  }
 }
