+++ frontend/src/components/NoteEditorForm.module.scss
@@ -1,0 +1,325 @@
+/**
+ * Shared note editor form styles
+ * Used by both the create page (notes/new) and edit mode (notes/[id])
+ */
+
+@use "@/styles/design-tokens" as *;
+@use "@/styles/mixins" as *;
+
+// ===== TIP BOX =====
+.tipBox {
+  padding: $spacing-4;
+  border-radius: $border-radius-md;
+  border: $border-width-thin solid $color-border-default;
+  background-color: $color-surface-tertiary;
+  margin-bottom: $spacing-4;
+
+  .tipHeader {
+    @include flex-between;
+
+    .tipTitle {
+      display: inline-flex;
+      align-items: center;
+      gap: $spacing-1;
+      @include text-secondary;
+      font-weight: $font-weight-semibold;
+      color: $color-text-heading;
+      margin-bottom: $spacing-1;
+    }
+
+    .dismissButton {
+      margin-left: $spacing-4;
+      color: $color-text-tertiary;
+      cursor: pointer;
+      background: none;
+      border: none;
+      padding: 0;
+      transition: color 0.2s ease;
+
+      &:hover {
+        color: $color-link-default;
+      }
+    }
+  }
+
+  .tipContent {
+    @include text-secondary;
+    color: $color-text-primary;
+  }
+
+  .tipExamples {
+    @include text-tertiary;
+    color: $color-text-secondary;
+
+    .tipAvoid,
+    .tipBetter {
+      display: inline-flex;
+      align-items: center;
+      gap: 2px;
+      font-weight: $font-weight-medium;
+    }
+
+    .tipAvoid {
+      color: $color-error-text;
+    }
+
+    .tipBetter {
+      color: $color-success-text;
+    }
+  }
+}
+
+// ===== ERROR =====
+.errorBox {
+  padding: $spacing-4;
+  border-radius: $border-radius-md;
+  border: $border-width-thin solid rgba(220, 38, 38, 0.3);
+  background-color: rgba(254, 226, 226, 0.8);
+  margin-bottom: $spacing-4;
+
+  .errorMessage {
+    color: $color-error-border;
+  }
+}
+
+// ===== EDITOR GRID =====
+.editorGrid {
+  display: grid;
+  gap: $spacing-6;
+
+  &.withSidebar {
+    @media (min-width: 1024px) {
+      grid-template-columns: 2fr 1fr;
+    }
+  }
+}
+
+.editorColumn {
+  @include stack($spacing-4);
+}
+
+// Template selector row - compact button placement
+.templateRow {
+  margin-bottom: $spacing-2;
+}
+
+// ===== FORM FIELDS =====
+.formLabel {
+  display: block;
+  margin-bottom: $spacing-1;
+  font-size: $font-size-sm;
+  font-weight: $font-weight-medium;
+  color: $color-text-secondary;
+}
+
+.formInput {
+  width: 100%;
+  padding: $spacing-2 $spacing-3;
+  border: $border-default;
+  border-radius: $border-radius-button;
+  background-color: $white;
+  transition: all 0.2s ease;
+
+  &:focus {
+    outline: none;
+    border-color: $color-input-border-focus;
+    box-shadow: $shadow-focus-primary;
+  }
+
+  &::placeholder {
+    color: $color-input-placeholder;
+  }
+}
+
+.formHint {
+  margin-top: $spacing-1;
+  font-size: $font-size-xs;
+  color: $color-text-tertiary;
+}
+
+// ===== REFERENCE TOGGLE =====
+.referenceToggle {
+  margin-top: $spacing-2;
+}
+
+.checkboxLabel {
+  display: inline-flex;
+  align-items: center;
+  gap: $spacing-2;
+  cursor: pointer;
+
+  .checkbox {
+    width: 16px;
+    height: 16px;
+    accent-color: $color-link-default;
+  }
+
+  .checkboxText {
+    font-size: $font-size-sm;
+    font-weight: $font-weight-medium;
+    color: $color-text-primary;
+  }
+}
+
+// ===== CHARACTER COUNT =====
+.charCountWrapper {
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  margin-bottom: $spacing-2;
+}
+
+.charCount {
+  text-align: right;
+  font-size: $font-size-xs;
+  color: $color-text-tertiary;
+
+  .count {
+    &.good {
+      font-weight: $font-weight-medium;
+      color: $color-success-text;
+    }
+
+    &.warning {
+      color: $color-warning-text;
+    }
+
+    &.error {
+      color: $orange-600;
+    }
+  }
+
+  .hint {
+    margin-left: $spacing-2;
+    color: $color-text-tertiary;
+  }
+}
+
+// ===== ACTIONS =====
+.actions {
+  @include flex-start;
+  gap: $spacing-3;
+  flex-wrap: wrap;
+  padding-top: $spacing-4;
+
+  .button {
+    padding: $spacing-2 $spacing-6;
+    border-radius: $border-radius-button;
+    font-weight: $font-weight-medium;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    border: none;
+
+    &:disabled {
+      cursor: not-allowed;
+      opacity: 0.5;
+    }
+  }
+
+  .saveButton {
+    background-color: $color-link-default;
+    color: white;
+
+    &:hover:not(:disabled) {
+      background-color: $color-text-secondary;
+    }
+  }
+
+  .cancelButton {
+    background-color: transparent;
+    color: $color-text-primary;
+    border: $border-width-thin solid $color-border;
+
+    &:hover:not(:disabled) {
+      background-color: $color-bg-secondary;
+    }
+  }
+
+  .aiButton {
+    background-color: $color-surface-tertiary;
+    color: $color-link-default;
+    border: $border-width-thin solid $color-border-default;
+
+    &:hover:not(:disabled) {
+      background-color: $color-surface-hover;
+    }
+  }
+}
+
+// ===== MODALS =====
+.modalOverlay {
+  position: fixed;
+  inset: 0;
+  z-index: 50;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: rgba($black, 0.5);
+}
+
+.modalContent {
+  max-width: 28rem;
+  padding: $spacing-6;
+  border-radius: $border-radius-lg;
+  background-color: $white;
+  box-shadow: $shadow-card-lg;
+}
+
+.modalTitle {
+  margin-bottom: $spacing-3;
+  font-size: $font-size-lg;
+  font-weight: $font-weight-semibold;
+  color: $color-text-primary;
+}
+
+.modalText {
+  margin-bottom: $spacing-4;
+  color: $color-text-secondary;
+}
+
+.modalList {
+  margin-bottom: $spacing-6;
+  padding-left: $spacing-5;
+  list-style-type: disc;
+  list-style-position: inside;
+
+  li {
+    margin-bottom: $spacing-1;
+    font-size: $font-size-sm;
+    color: $color-text-secondary;
+  }
+}
+
+.modalActions {
+  display: flex;
+  gap: $spacing-3;
+
+  .modalButton {
+    flex: 1;
+    padding: $spacing-2 $spacing-4;
+    border-radius: $border-radius-button;
+    font-weight: $font-weight-medium;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    border: none;
+
+    &.primary {
+      background-color: $color-link-default;
+      color: $white;
+
+      &:hover {
+        background-color: $color-text-secondary;
+      }
+    }
+
+    &.secondary {
+      background-color: transparent;
+      color: $color-text-secondary;
+      border: $border-width-thin solid $color-border;
+
+      &:hover {
+        background-color: $color-bg-secondary;
+      }
+    }
+  }
+}
